requirements.txt
faiss-cpu==1.9.0
fastapi==0.115.11
numpy==2.2.3
python-dotenv==1.0.1
Requests==2.32.3
sentence_transformers==3.4.1
streamlit==1.43.0
